Stable

+++ app/src/main/java/com/evolgames/entities/commandtemplate/commands/JointDestructionCommand.java
@@ -21,16 +21,17 @@ public class JointDestructionCommand extends Command {
         Log.e("Mirror", joint + "-------------Destroying joint-----------:" + joint.getType());
 
         PhysicsWorld physicsWorld = Invoker.scene.getPhysicsWorld();
-
         JointBlock jointBlock = (JointBlock) joint.getUserData();
-        jointBlock.getEntity().removeJointBlock(jointBlock);
-        jointBlock.getBrother().getEntity().removeJointBlock(jointBlock.getBrother());
-        jointBlock.setAborted(true);
-        jointBlock.getBrother().setAborted(true);
-        if (this.joint instanceof MouseJoint) {
-            Invoker.scene.onDestroyMouseJoint((MouseJoint) joint);
+        if(jointBlock.isNotAborted()) {
+            jointBlock.getEntity().removeJointBlock(jointBlock);
+            jointBlock.getBrother().getEntity().removeJointBlock(jointBlock.getBrother());
+            jointBlock.setAborted(true);
+            jointBlock.getBrother().setAborted(true);
+            if (this.joint instanceof MouseJoint) {
+                Invoker.scene.onDestroyMouseJoint((MouseJoint) joint);
+            }
+            physicsWorld.destroyJoint(joint);
         }
-        physicsWorld.destroyJoint(joint);
     }
 
     @Override
